refactor(update-weight): group protocols/ and hf_weight_iterator/ packages
protocols/{broadcast,delta,p2p,p2p_transfer_utils,rdt,cuda_ipc}.py (colocate
renamed cuda_ipc after its transport); hf_weight_iterator/{__init__,bucketing,
atomic_groups}.py. The iterator package keeps its import path via __init__.
