+++ tests/fast/backends/training_utils/weight_update/test_atomic_groups.py
@@ -5,7 +5,7 @@ from tests.ci.ci_register import register_cpu_ci
 register_cpu_ci(est_time=60, suite="stage-a-cpu", labels=[])
 
 
-from miles.backends.training_utils.weight_update.atomic_groups import get_hf_atomic_update_groups
+from miles.backends.training_utils.weight_update.hf_weight_iterator.atomic_groups import get_hf_atomic_update_groups
 
 
 class TestHfAtomicUpdateGroups:
